Refactor MyInfo controllers and update routes for improved organization
